cast

+++ platform/platform-impl/src/com/intellij/openapi/updateSettings/impl/pluginsAdvertisement/PluginsAdvertiser.java
@@ -109,7 +109,7 @@ public class PluginsAdvertiser implements StartupActivity {
                   }
                 }
               }
-              indicator.setFraction(idx++ / unknownFeatures.size());
+              indicator.setFraction(((double) idx++) / unknownFeatures.size());
             }
 
             try {
